backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
langgraph==0.0.20
langchain==0.1.0
langchain-core==0.1.10
python-dotenv==1.0.0
kiteconnect==4.2.0
openai>=1.12.0
anthropic>=0.18.0


